requirements.txt
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
langchain-core==1.2.6
langchain-google-genai==4.1.2
passlib==1.7.4
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
SQLAlchemy==2.0.45
uvicorn==0.40.0

